requirements.txt
tensorflow>=2.12
scikit-learn>=1.3
torch>=2.1
transformers>=4.43
tqdm>=4.66
numpy>=1.24
pillow>=10.0
requests>=2.31